afisha/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
django-tinymce==5.0.0
dotenv==0.9.9
isort==7.0.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
